requirements.txt
streamlit==1.32.0
nltk==3.8.1
scikit-learn==1.4.1.post1
requests==2.31.0
numpy==1.26.4
pandas==2.2.1
python-dotenv==1.0.1
newsapi-python==0.2.7 